showing images in two options, cells view or rows view
the user can choose to show the images in cells view to see n x n table
with an image in every cell.
Or rows view to see 1 x n table with an image in every row as used
before. The preference is persisted in the config file with the tag
images_view